refactor: merge jwt-signer-spi into jwt-signer
