+++ core/common/lib/crypto-common-lib/build.gradle.kts
@@ -19,7 +19,7 @@ dependencies {
     api(project(":spi:common:keys-spi"))
     api(libs.nimbus.jwt) // nimbus classes are exposed on the API surface of CryptoConverter and DefaultJwsSignerProvider
     implementation(project(":spi:common:boot-spi"))
-    implementation(project(":spi:common:jwt-signer-spi"))
+    implementation(project(":spi:common:jwt-spi"))
 
     // used for the Ed25519 Verifier in conjunction with OctetKeyPairs (OKP)
     runtimeOnly(libs.tink)
