+++ core/common/token-core/src/main/java/org/eclipse/edc/token/TokenServicesExtension.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.edc.token;
 
-import org.eclipse.edc.jwt.signer.spi.JwsSignerProvider;
+import org.eclipse.edc.jwt.spi.signer.JwsSignerProvider;
 import org.eclipse.edc.jwt.validation.jti.JtiValidationStore;
 import org.eclipse.edc.keys.spi.PrivateKeyResolver;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
