+++ settings.gradle.kts
@@ -296,7 +296,6 @@ include(":spi:common:keys-spi")
 include(":spi:common:identity-did-spi")
 include(":spi:common:json-ld-spi")
 include(":spi:common:jwt-spi")
-include(":spi:common:jwt-signer-spi")
 include(":core:common:junit-base")
 include(":spi:common:token-spi")
 include(":spi:common:oauth2-spi")
