--- spi/common/jwt-signer-spi/build.gradle.kts
@@ -1,26 +1,0 @@
-/*
- *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-
-plugins {
-    `java-library`
-    `maven-publish`
-}
-
-dependencies {
-    api(libs.nimbus.jwt)
-    api(project(":spi:common:boot-spi")) //Result
-    implementation(libs.edc.runtime.metamodel)
-}
-
-
