--- spi/common/jwt-signer-spi/src/main/java/org/eclipse/edc/jwt/signer/spi/package-info.java
@@ -1,18 +1,0 @@
-/*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
- *
- *  This program and the accompanying materials are made available under the
- *  terms of the Apache License, Version 2.0 which is available at
- *  https://www.apache.org/licenses/LICENSE-2.0
- *
- *  SPDX-License-Identifier: Apache-2.0
- *
- *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
- *
- */
-@Spi(value = "Implementation SPI that is used to contribute custom JWSSigners to the JwtGenerationService")
-
-package org.eclipse.edc.jwt.signer.spi;
-
-import org.eclipse.edc.runtime.metamodel.annotation.Spi;
