+++ spi/common/jwt-spi/src/main/java/org/eclipse/edc/jwt/spi/signer/JwsSignerProvider.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *  Copyright (c) 2026 Think-it GmbH
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,11 +8,11 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *       Think-it GmbH - initial API and implementation
  *
  */
 
-package org.eclipse.edc.jwt.signer.spi;
+package org.eclipse.edc.jwt.spi.signer;
 
 import com.nimbusds.jose.JWSSigner;
 import org.eclipse.edc.runtime.metamodel.annotation.ExtensionPoint;
